pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "email-validator>=2.2.0",
    "flask-sock>=0.7.0",
    "flask>=3.1.1",
    "google-cloud-texttospeech>=2.27.0",
    "gunicorn>=23.0.0",
    "openai>=1.93.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "twilio>=9.6.4",
    "websockets>=15.0.1",
    "werkzeug>=3.1.3",
    "gtts>=2.5.4",
    "requests>=2.32.4",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "sift-stack-py>=0.7.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.175.0",
    "google-auth>=2.40.3",
    "schedule>=1.2.2",
    "reportlab>=4.4.2",
    "pillow>=11.3.0",
    "flask-mail>=0.10.0",
    "python-dateutil>=2.9.0.post0",
    "qrcode[pil]>=8.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "pydub>=0.25.1",
]
